feat: Add streaming analysis support and improve configuration
- Add streaming analysis endpoint with Server-Sent Events
- Implement CORS middleware for cross-origin requests
- Add progress tracking and real-time updates in frontend
- Update Ollama configuration for Docker environment
- Increase max prompt length for better analysis quality
- Improve error handling and user experience

+++ backend/explorer.go
@@ -111,7 +111,7 @@ func readFileContent(absFilepath string) (string, error) {
 
 	// Lire le contenu du fichier
 	size := fileInfo.Size()
-	maxSize := config.AppConfig.Analysis.MaxFileReadSize
+	maxSize := int64(config.AppConfig.Analysis.MaxFileReadSize)
 	if size > maxSize {
 		logrus.Warnf("File '%s' (%d bytes) is too large. Reading partially.", filepath.Base(absFilepath), size)
 		content, err := os.ReadFile(absFilepath)
